[debugger] Convert variables to expression values.
This is a first pass at a class to convert DWARF Variables to the
expression value objects used in the debugger expression evaluator.

The converter currently assumes everything is a [u]int64_t or smaller
integer stored directly in a register. Anything more complicated will
require asynchronous memory requests to dereference pointers which will
be done in a separate pass.

A symbol data provider is also added to the Frame for use in doing these
evaluations in production, but this is not completely hooked up yet.

The symbol data provider is now reference counted to deal with lifetime
issues with respect to its associated frame.

A LazySymbol no longer asserts if one attempts to dereference it while
it's not valid. This used to be the behavior but I changed the spec in a
previous patch without updating the implementation.

The MockSymbolDataProvider was extracted from the DWARF expr eval
unittest into its own header so it can be shared.

TEST=unit

Change-Id: I5dfe93263e134c3d92a68f3a0fd953a10833f904