[audio] Add FxLoader for effect SO
This CL adds the FxLoader class, to encapsulate the dynamic
loading and import of the audio effect SO library. This
singleton object will serve as the central point of focus
for communication to the loaded library. This CL also adds
tests that use the actual fx_loader, so the test binary
can not only fully test our loader, but also test a .SO
library exactly as it will be used when officially loaded
(by the same fx_loader) into the audio_core process.

Tests: added relevant tests to audio_dfx_tests. For now,
no behavior change (no effect SO is loaded by default).

Change-Id: Ia287995008fad0b0ad482853f7a711b93d5faac2