[kernel] Pull MSI support out of PCIe into dev/interrupt
- Remove PCI dependency from the Intel IOMMU impl.
- Remove WITH_DEV_PCIE defines that gate MSI support.
- Move msi_* functions to dev/interrupt.h and provide
  weak stubs in dev/interrupt/msi.c
- Update intel and aram methods to use new MSI functions
  that exist standalone.

The MSI interface is subject to change still in the future,
but for now this decouples the IOMMU from PCI and sets up
a framework for MSI being part of the kernel interrupt interface.

PciePlatform support classes are mostly left alone to avoid
refactoring a system that will be unused later. They may be
eliminated in a future CL if it seems easy to do.

Test: core-tests and boot tests

Change-Id: I89f226ccf5ebf4930ea06d9f32cb6e1caabde34e