[audio] Add FxProcessor
This CL adds the FxProcessor object, a companion to the
previously-added FxLoader. The FxProcessor encapsulates
an ordered chain of instantiated effects, to be associated
with an AudioOutput or AudioInput. This CL does not go so
far as to make that association -- it just adds the class
itself, plus tests to verify it in isolation.

Test: additional tests in audio_dfx_tests. Still no change
in behavior.

Change-Id: I5574af4c5e2022b97c683353fe4d867547491f42