[netstack] Refactoring the type of ios.dataHandle
The data socket handle was created as a zx.Socket, but casted to
a zx.Handle before stored in ios.dataHandle. This caused ios.dataHandle
to be cast back to a zx.Socket in several places before used.

This CL changes the type of ios.dataHandle to zx.Socket so  avoids
this double type casts.

Test: netstack_loopback_test, CQ

Change-Id: I093046602ecca9439d9bda84701c3e410165cc80