[fidl] Add source information to AST
Every node in the AST is given start, end, and previous_end
information. The start and end are relatively self-explanatory.  The
previous_end is the end of the last AST node in declaration order.  The
space between the previous_end and start will most likely contain
any comments relevant to the node.

This may leave the AST with a trail of stuff (whitespace, comments) that
is stored in the top level AST node (File).

Test: A visitor will be added in a subsequent CL.  It will have a test
that ensures you can reconstruct the original file.

Change-Id: I3f7dcd7d8ca124042679b1b01e8dd0cada270825