[component] Component inspection interface.
This change defines a new inspection interface for components consisting
of the following:

|Object| - A class that describes some "Object" of interest to the
component. Consists of low-level interfaces to set string "Properties"
and numeric "Metrics" on the object. Also has interfaces to set Object
children on a Object. Implments an custom FIDL interface as well as
LazyDir to expose the values to interested parties.

|ObjectDir| - A wrapper around |Object| that adds convenience
getters/setters and hierarchy traversal functionality. ObjectDir is
exposed as part of components' outgoing/ directory in the hub.

|ExposedObject| - A base class for any class in a component that makes
sense to expose as an |Object|. Automatically handles parent-child
relationships and cleanup on object destruction.

TEST=follow up
CP-59: #progress

Change-Id: I3223b3f5563bba05a45afe6e520506c72bf79954