[vulkan] Add image_pipe_swapchain layer to SDK
This adds the VkLayer_image_pipe_swapchain layer to the SDK alongside
the core validation layers. The image pipe swapchain layer currently
dynamically links against libasync-default.so, which is slightly
unfortunate as it should be hermetic, but it doesn't actually use the
library at all at runtime so it's at least somewhat usable by clients.

MA-475 #comment
Test: Build SDK and see libVkLayer_image_pipe_swapchain.so and json in
the expected locations

Change-Id: Icd0182077defb46c39ddf9ec977401e6af33fcb3