[scenic] Add CustomFocusBehavior to viewsv1::ViewProperties
Soft keyboards need the ability to receive pointer events without
stealing away keyboard focus. This change adds a new, optional struct
to ViewProperties, that allows a view to ban focusing of it or any of
its children. We opted to add a struct instead of just a raw bool on the
ViewProperty so that it could be made optional (and
backwards-compatible), with a default behavior of allowing focus.

This change would likely be superseded by a FocusEvent in the new
Views, but it'd be nice to get this temporary solution right now so the
IME team can experiment with soft keyboards.

This change also adds a --disable-focus option to Tiles to test out the
new behavior;  You can quickly check out the behavior by building
ermine, and then running `tiles_ctl add latin-ime` followed by
`tiles_ctl add --disable-focus latin-ime`. You should get two
keyboards that both accept touch input, but only one should
steal keyboard focus away from the terminal.

TXT-11 #done

SCN-164 #comment Still likely are some bugs, I think re: a pointer
touch event not sending a cancel event to the previously focused view,
so I think it shouldn't be closed yet. That said, this CL should improve
behavior for multi-finger input.

Test: Manually tested focus behavior with the newly added flag
to tiles_ctl.

Change-Id: I20bb595bea152502040da26f70015ce6b2a729eb