[runtests-utils] New verbosity arg handling
The old code had too small a buffer, since the widest string from a
signed char is "-127". Rather than just fixing that, use strings and
vectors to avoid the issue altogether.

The "should we add a verbosity arg" logic has confused me for
years, so hopefully it's more clear now.

Also, clang-format hit a couple of existing lines in the file.

Tested:
Ran "runtests" and "runtests -v" in QEMU and verified with a printf
that the expected args made their way to FuchsiaRunTest().

Change-Id: I74a85c7f14fdf77fa60e24b0eac72f2aeda75a4c