[Codec] stop dropping CodecFactory requests
Avoid pumping the async::Loop between CreateFromStartupInfo() and
AddServiceForName(), because doing so was allowing incoming requests for
a CodecFactory to be dropped.

Also fix up some related test code to be on a binding's dispatcher when
making FIDL calls related to that binding.

Tested: use_h264_decoder_test + amlogic_video_decoder tests + CQ
Change-Id: I700c4bee097a4b4a8549e0809e17b040363042a5