[build] Switch to C++ 17 (reland)
This is a reland of 08c8dfac721cd3c18b4e5bb484e7d7b46f9ca215

Building the Zircon layer with -std=c++17 ensures that the code
(to the extent it is built/tested by Zircon bots) stays compatible
with C++ 17.  Garnet and higher layers still build with -std=c++14
to catch any regressions in compatibility with C++ 14 so roller
failures will push back for reverts or fixes in Zircon as needed.
It will eventually be the responsibility of the SDK build to do
compatibility testing for the APIs exported to the SDK and push back
for reverts or fixes in layers.

The placement new declarations need noexcept in C++ 17 mode and
it has no effect in C++ 14 mode.

Bug: BLD-227 #comment Zircon uses -std=c++17
Test: CQ
Change-Id: I93936b81ad154db1798c29b4d37d506176f7bea9