[debugger] Add "print" command.
Adds an ExprEvalContext implementation that provides a bridge between
the expression subsystem and the symbol subsystem. A Frame provides this
implementation to evaluate expressions in the context of that stack
frame.

Long frame formatting now uses asynchronous output since it needs to
print the parameters which may be asynchronous to evaluate.

The FormatValueHelper which helps with async printing is now refcounted
which eliminates some nasty memory management around its use. It is now
scoped to its own callback.

Options for formatting are added. These are not currently hooked up to
the UI but there are tests for them. These allow, for example, hex
printing of integers.

Better error messages are hooked up to tokenizing and parsing that
provide context for where the error occurred.

TEST=unit

Change-Id: I5a548238fc082229cfa3d9fe473e548bc552e65b