[dev][platform-bus] Remove PDEV_ADD_DISABLED and pbus_device_enable()
These allowed adding platform devices in a disabled state,
meaning that they did not actually exist in the devmgr
until they were later enabled by calling pbus_device_enable().

Removing this since it is no longer being used and adds additional
complexity that is not being tested.
If we need something like this in the future, I think it would be
better to add a pbus_device_remove() API instead.

TEST: manual testing on qemu and hikey960
Change-Id: I3bb5b48116bd81d3a48ea12801cfec56c19775c4