[fzl] Introduce ResizeableVmoMapper
Introduce ResizeableVmoMapper; an extension of the existing VmoMapper
which allows resize operations.  This is intended to replace the
existing MappedVmo class in an attempt to unify/condense our C++ VMO
utility code.

Existing unit tests were ported, and new tests added for the small
number of feature changes.  In specific, ResizeableVmoMapper can be
instantiated as an embedded member of an object (or on the stack) and
moved around in addition to being a heap allocated object managed with
a unique_ptr.

ZX-2462 #comment VmoMapper and MappedVmo functionality is now unified

Test: Build and unit tests.
Change-Id: I89c07db20bd8864a7e68a2f0061f7adea2ebc56c