[display] Remove ddk calls from on_displays_changed
This change replaces get_info with per-display argument structs to
on_displays_changed, and it moves the display mode validation to a
posted task. This allow simpler locking in the impl drivers, and it
enables the addition of output parameters into the argument structs.

This change also causes ownership messages to be sent before display
changed messages. Update various clients to not rely on a particular
order of messages.

Test: Boot to virtcon/hotplug on NUC/vim/eve. Run gfxtest, gfxlatency,
and display-test

Change-Id: Id561ed7587037353c1aeeab96511481d74026bdc