[tee][optee] Support private shared memory RPCs
The OP-TEE driver is responsible for mediating communication between the
REE and TEE. During this communication, the TEE may request the driver
to send messages to and perform commands in the REE on its behalf via
RPCs, but in order to communicate these requests to the driver, the TEE
needs support from the driver to marshal shared memory (between the
driver and TEE) via an RPC call. This CL adds support in the OP-TEE
driver for responding to these RPCs and managing shared memory from the
private pool for driver-TEE communication.

Test: Manual testing on VIM2 to verify shared memory RPCs are properly
fulfilled and communicated back to the TEE.

Change-Id: Ib5d66b045d43fefdff20536a2e6bdd02e53217d4