[audio][core] Pass-thru effect SO
This CL adds a small libaudiofx.so library, containing a simple
pass-thru effect that consumes no CPU and incurs no latency.

This CL also adds the Load and Unload of this library, done by
the AudioDeviceManager singleton in its Init and Shutdown.

Subsequent CLs will relate the available effects in the loaded fx
library, with the AudioDeviceSettings for that device which instruct
us how to configure any effects for that device.

Test: audio_dfx_tests passes on all archs

Change-Id: Iac14250826885a322d07a58c728e358b6c8d05db