[pm] Add delta command to diff package snapshots
This change adds a new pm command to compare 2 package snapshot files.
With no extra parameters (other than the paths to the two snapshot
files), pm delta will display summary information, as well as detailed
information about packages and blobs with the largest download size.
There are various options to customize this human readable output, or,
if an "--output (<path>|-)" option is provided, pm delta will emit the
statistics as JSON to the provided path or stdout.

PKG-175 #comment

Test=manual

Change-Id: I9aaf9318d3573acf7aa5fb8517246312e336132b