[zircon][sdio] Use sdio_rw_direct command whenever possible
When using sdio_io_rw_extended command with single byte reads, the broadcomm
chipset does not complete the read request and does not raise an interrupt.
Neither does it fail the command with an error. This changeset uses the
sdio_rw_direct command for 1-byte reads to avoid this situation.

Test: Run brcmfmac driver on vim2. The driver made more progress. The
reading of the func1 misc register is not stuck anymore.

Change-Id: If38a6b9d3ee6d0b0d6f0f4fc4dfd6babe8553521