[devmgr] Fix path walk bug across local nodes
Certain nodes in devfs are considered 'local', meaning
that they should not hand off 'Open' RPCs to the underlying
devhost. While path walking to these nodes, if additional
path components exist (in the case of ZX_ERR_NEXT),
an error should be returned. The prior behavior
instead opened the local node, silently discarding the
remainder of the path.

This means that although the following will open
successfully:

  open("/dev/sys/pci/00:01.0")

and

  open("/dev/sys/pci/00:01.0/bochs_vbe")

(Since "/dev/sys/pci/00:01.0" is a local node)
The following will ALSO open successfully:

  open("/dev/sys/pci/00:01.0/this_does_not_exist")

Test: `ls /dev/sys/pci/00:01.0/bochs_vbe_but_not_real`

Change-Id: I868e4159c27502a0d784e7c3c624820a8b645507