[machina][x86] Simplify MMIO addresses
Move all devices to addresses between 0xf8000000 to 0xf8ffffff. On x86,
we have to keep all of these addresses under 0xffffffff as the MADT can
only hold a 32-bit address for the IOAPIC address, and both Zircon and
Linux have restrictions on where the PCI addresses can exist.

This change also removes the TPM handler, as it's no longer needed.

MAC-151 #comment

Test: Ran "machina_unittests", "zircon_guest", and "debian_guest".
Change-Id: I0f6198e49d9fdf653670d7f2d67ad42179f4bebf