[lib/ui] Move scenic impl into scenic_impl namespace
Previously, the client library and the server were in the same
namespace, which was awkward because they define some of the same
symbols. To make that work, we needed to be careful never to link the
client and server together, as sometimes happens in tests.

This CL moves the scenic server into the scenic_impl namespace,
separating it from the client library. This change makes it possible to
change the client library to be a single gn target (and therefore
possible to include in the SDK as a single library) without violating
ODR in scenic's tests.

Test: No behavior change.

Change-Id: I1876ceebb5ca58e921a1a3716072099b64f5e153