[mediaplayer] fuchsia.mediplayer readability
This CL updates the fuchsia.mediplayer FIDL definition to be
compliant with the readability rubric and otherwise be closer to
the recent API proposal. This includes:

- renaming MediaPlayer to Player
- taking initial steps to delegate advanced source and sink features
  to the SourceManager and SinkManager interfaces
- removing the MediaResult enum

For now, SourceManager and SinkManager interfaces are replaced by two
deprecated methods to make this CL simpler. Clients other than
mediaplayer_tests don't currently use these methods.

The SourceManager definition is included but not currently used. The
SinkManager definition in this CL is largely a placeholder...it has
yet to be completely defined.

TEST: Existing integration and unit tests, manual playback tests.

Change-Id: Idfdeda26c1573a41b9512022580a1dd87865fd2b