[bt][l2cap] Fix ConfigurationResponse ID parsing
In BrEdrCommandHandler, the response parsing for outbound Configuration
Requests processed the "Source Channel ID" field as ID for the remote
peer's endpoint when it should be the ID for the endpoint on the
"source" (i.e. requester). The ConfigurationResponder for inbound
Configuration Request handling already handles the ID direction
correctly.

Fix implementation and test.

NET-704 #comment

Tested: in bt-host-unittest, update
        L2CAP_BrEdrCommandHandlerTest.OutboundConfigReqRspPendingEmpty

Change-Id: I487cddf9248ec2562515c40b8929063b4b2529df