[bt][sdp] Connect SDP server to L2CAP PSM
Connect the SDP server to the L2CAP PSM, meaning it can
answer queries from remote devices that connect to that PSM.

Update the ServerTest to use a L2CAP::FakeLayer so we can test
incoming connections being routed correctly.

This is a reland of dbb7e62958b9d0fea56ff85def9d6f31010fca5b with
a fix to turn a crash into a warning for unimplemented
L2CAP::RegisterService.

NET-1007 #comment
Test: SDP_ServerTest.*

Change-Id: I24ef20f84e24529cfc400f01abaa8966835f077e