[svchost] restore MediaPlayer to svchost.cpp
https://fuchsia-review.googlesource.com/c/garnet/+/184391 will
replace fuchsia.mediaplayer.MediaPlayer with
fuchsia.mediaplayer.Player.
https://fuchsia-review.googlesource.com/c/zircon/+/184726 was
submitted to change the name in svchost.cpp. The test that
requires this is still not working, so this CL lists both the
old a new names to unblock the zircon roller and enable
debugging work to proceed. MediaPlayer will be removed later.

TEST: ran mediaplayer_tests

Change-Id: I074212855594875879b901e61148f16a9a0c1723