[debugger] Add support for array types.
This adds support for the DWARF array type.

Support is added for array dereferencing in expressions so that
"print a[4]" will work assuming "a" is an array or a pointer.

The ExprNode (node generated by the parser) is now reference counted.
This was necessary to keep things in scope dring asynchronous
evaluation. There are a lot of copy-paste changes associated with this
but they are simple.

There is still no support for printing arrays.

TEST=unit

Change-Id: Id57638dde215ae2e49eec9f6ad24e521b015ec18