[hypervisor] Fix Trap destruction
Remove packets from the port dispatcher queue on destruction of the Trap
to prevent a bug where packets are freed after the arena that contains
them has been destructed.

Since Trap packets have nullptr as the handle a small modification to
CancelQueued was needed to allow PortPackets with null handles to be
canceled.

TEST=Run guest_integration_tests

Change-Id: Ibb394bf5e123bfac78942eb58d6c7d4acb1a939f