[intents] Add IntentHandler support to StoryController
This forwards intents to a module's IntentHandler interface, if such an
interface is exposed. The framework will forward all Intents it receives
for a particular module, regardless if the module is already running or
not.

An integration test is added which makes sure the intent is forwarded
the first time a module is started, on a subsequent intent which does
not restart the module, as well as a subsequent intent which does
restart the module.

This change does not transform the intent parameters to use link names
which are sensible to the handler, and it does not remove the old intent
parameter delivery method (i.e. setting the links with the given
parameter names).

TESTED=tests/intents
MI4-1246 #done

Change-Id: Ib93b320fec0f74d3a3369b177548803a58c00a53