[symbolize] Check modtime on ids.txt
This change checks the mod time on the ids.txt file
prior to reading the file and thus prior to performing
verificiation of the ids.txt. Coupled with caching this
reduces the primary cost of the symbolize to be parsing (30-40%, 30% regex alone),
lock contention (15%). This represents something like a 1000x
speed up. Further optimization seems much harder. Increasing
go channel buffer sizes should give a 25%-50% improvement.

Change-Id: I0da1a05c7e31afc68d8ddc2eecd0acf42ae59961