[ledger] sandbox benchmarks
This patch enables the execution sandbox for Ledger benchmark binaries,
adding a CMX manifest for each binary and changing the invocation path
(binaries referred to through the pkgfs full path run without the
sandbox).

In the next step we will switch the sync benchmarks to embed the
credentials file in the package and drop the "system-temp" permission.

LE-557 #comment [ledger] sandbox benchmarks

Test: run the benchmarks
Change-Id: Ic7b217135d177e147b8a2c2e42e2d45600ca6c7b