[appmgr] Don't provide backing_dir to root realm.
Currently, appmgr hosts fuchsia.sys.Loader, whose channel is passed as
the backing dir for the root realm. Thus there is a self-referential
dependency between appmgr and itself. Currently this is not a problem,
but soon to support Readdir appmgr will be sending a readdir() call to
itself. If we kept the backing dir, the root realm's readdir() to itself
would deadlock.

CP-23 #progress

TESTED=ran garnet tests

Change-Id: Ibfcb398eee793a3cc4d204128a259f7f1b7bc6ec