[fs-test-utils] Fix test suite name outputted in the JSON perf results file
Before, the test name was copied as the test suite name, which is not
the right thing to do.  An example is
"minfs/Bigfile/16Kbytes/1024-Ops/1-Cycle/Write".  When uploaded to the
Catapult dashboard, this gets rejected with a "400 Bad Request" error,
presumably because of the slashes in the name.

This should fix the garnet-x64-perf-swift_canyon bot, which is failing
because its Catapult upload step now fails.

Bug: ZX-2492
Test: "/boot/test/fs/fs-bench-test --out /tmp/results.json && dd if=/tmp/results.json"
  and check output
Change-Id: Idd0644ee37e89def020f79f8aa1e6c9fd2f1d490