[debugger] Add a utility for reading arrays.
Add some utility functions that resolve array types to vectors or
values. These are not yet used in production code.

This includes some changes to memory reading to better handle invalid
memory.

Did some cleanup in the SymbolVariableResolver to avoid misuse by
asserting if more than one thing is requested simultaneously and to
guarantee that the callback is not issued if the resolver is destroyed.

A test harness is added that sets up a message loop. This is necessary
for some types of tests.

TEST=unit

Change-Id: Ifcb1bf87b924ffc8a269dd502ecd8b2816918b98