[sdk] Start generating SDK tarballs directly from the build.
This change introduces a new build artifact for sdk targets. A tarball is created with the contents of every atom in the SDK.
Currently the file layout in the archive is non-sense, as atoms specify their contents relatively to the atom root in the SDK,
whereas the path to generate the archive expects these paths to be relative to the SDK root.

This change also introduces metadata for both SDK atoms and the SDK in general. That metadata
will be used to drive the integration of the SDK into foreign build systems.
Currently only sdk_executable targets produce metadata. Support for other targets will be added in
subsequent changes.

Test: build produced an SDK tarball with metadata
Bug: DX-340

Change-Id: I424d068a5dd8fbdd35c26b6d1dbd8053e6672801