[kernel] Fix memory leak during thread_exit
A detached thread's stack and/or struct are freed using a deferred
procedure call (DPC). Queued DPCs are stored in a list using an
intrusive list_node.

When queuing a DPC, dpc_queue[_thread_locked], checks to see if the
DPC is already in a list somewhere by looking at its list_node field.
If the DPC is already in a list, then the call fails with
ZX_ERR_ALREADY_EXISTS.

Prior to this change, the DPC used to free a thread's resources was
never initialized. When queued, its list_node was stack garbage and so
the call would frequently fail resulting in the thread's stack and/or
thread_t being leaked.

Test: hacked up thread_tests to create a bunch of detached threads
that simply exit and used kstats -m to see that heap remained constant

ZX-2512 #done

ZX-2509 #comment pre-req (dpc queue)

Change-Id: I0873a83693f7a753edbdccf21f87442c0c48569e