[tee][optee] Implement basic RPC command handling
One of the core RPCs in the OP-TEE spec executes a command specified and
parameterized by an OP-TEE message format stored in memory. This CL
adds support for the commands that allocate/free shared memory and load
trusted applications for the trusted world. Supporting these commands
are enough to allow for simple trusted applications to be loaded and
executed in OP-TEE.

Test: Manual testing on both VIM2 and Astro to verify that the
TEEC_OpenSession API loads and returns successfully for a test "hello
world" trusted application-client application pair.

Change-Id: I1a2baa39c81090c678506e5b76abd792236835dd