[vim2] Relax clock restrictions on VIM2 HDMI audio
Earlier, we were disallowing audio over HDMI on VIM2 (AmLogic S912)
unless we were driving on of the standard CEA/CTA video pixel clock
rates.  In particular, we were only allowing audio if the pixel clock
rate was 25.2, 27, 54, 74.25 or 148.5 MHz.

This is mostly because we had a small display we were attempting to
work with (a GeChic OnLap 1002) which seemed to do OK with audio run
over standard clock rates, but was having what felt like audio clock
recovery trouble with some of the non-standard rates (including its
preferred pixel clock rate of 71MHz)

We have some other displays, however, which seem to do just fine with
both standard and non-standard pixel clock rates.  Additionally, it is
unclear what is making the GeChic display unhappy, since we are able
to compute valid N/CTS values for its nominal pixel clock rate (at
least for a 48KHz audio frame clock rate) without needing to perform
any dithering of N/CTS.

For now, just relax the restriction which was disallowing audio when
running with a non-CEA pixel clock rate.  We can track the issue with
the GeChic display with a bug, and (if the display becomes popular),
address the audio issue using a quirk higher up in the stack.

Test: manual on VIM2 with Eyoyo EM10K HDMI display
Change-Id: If38194c8bd5a10096aaa1c65ff4ebbf3f3ee8dd4