[nand] Add a nand device test driver.
This is a driver that binds to any nand protocol device and exposes an
ioctl interface to perform basic functionality tests.

The test application performs unit tests to the broker driver itself,
and can also run those tests over a nand device given as a command
line argument.

There are two basics mode of operation: bind directly on top of a nand
protocol driver (which means that said device cannot be already bound),
or use a broker device that has been bound already to a nand device.

In any case, modifications to the binding rules of other devices on
the system are most likely required to control what to test.

Test: nand-test over the broker, proxy, ram-nand and (raw-)nand drivers.
Change-Id: I51cb729d38d2d33fb610d4b106bb6b6587d0d8cb