[devmgr] Do not allow creating invisible device in create() callback
When a new devhost is created, the create() callback creates a new device
to act as a proxy for the corresponding device in the parent devhost.
At this point the device in the parent devhost is visible,
so the proxy device must be visible as well.
Currently DEVICE_ADD_VISIBLE is ignored when creating the proxy device.
This CL makes device_add() return ZX_ERR_INVALID_ARGS instead.

ZX-2511 #done

TEST: manual testing in qemu. Verified that device_add() returns error in this case.
Change-Id: Ia519a654c69356bae0f0d7464a64e022293de2d0