[amlogic][mailbox][scpi] Move mailbox and scpi out of platform bus devhost
Originally the Amlogic mailbox and SCPI drivers were loaded in the
root platform bus devhost to allow multiple platform devices to
use their protocols. This required having the platform bus proxy their
protocols, which required adding SOC specific protocol support to
the generic platform bus implementation.
However the aml-thermal driver is the only client of the SCPI protocol,
so the proxying is no longer necessary.

Now that we have a mechanism for providing the platform device protocol
to children of platform devices, we can now move them all into a separate
devhost and no longer rely on platform bus proxying the SCPI protocol.
Now the aml-mailbox driver is a standard platform device that runs in
its own devhost. aml-scpi runs as a child of aml-mailbox and vim-thermal
is a child of aml-scpi.

TEST: manual testing on VIM2. Thermal driver loads and initializes properly.

Change-Id: I281b4015b6be7953359aa6efbba702de9cded557