[bin/cobalt][bin/network_time] Serve or close incoming handles.
cobalt:
The async loop is created in this program, but it is never run. This
change utilizes the time the collection thread sleeps to run the message
loop. This has the result of serving the incoming directory request.

network_time:
This component doesn't have any output so serve, so we explicitly close
the handle so listings won't hang.

These changes prevent listing the hub directory from hanging.

TEST=`$ find /hub` now does not hang.

CP-123: #done
Change-Id: Iaa2db51bf557bee98885b0652be013273876cc14