[ime] [view_manager] Onscreen keyboard infrastructure
There are two parts to this change.

First, we've added a new InjectInput to the main ImeService. Previously,
you could only call InjectInput on individual IMEs corresponding with a
particular text field, which was fine, since Scenic knew which IME was
active. However, a onscreen keyboard does not know this, and needs a
central place to put in input. The input is always sent to the most
recently created IME.

Second, we've added new KeyboardShow() and KeyboardHide() methods on
ImeService. Scenic has been updated to call these when Flutter tells us
to, and there is a new ImeVisibilityService that onscreen keyboard
containers can access to know when to open or close the keyboard.

This change also switches IMEService to use a slightly different way of
definining service methods that is hopefully easier to read, especially
if a single data structure is hosted as two different services with
different methods. I'd like to eventually switch over the IME itself
too, but that's probably best for another change.

TEXT-19 #comment created technical debt

TEXT-22 #comment created technical debt

TEXT-13 #comment mostly done, just needs a few modifications to
PlatformView so focus from Flutter is properly sent all the way over.

TEXT-14 #done

TEXT-15 #done

Test: Manual, also wrote a small example program to manually validate
that ImeVisibilityService works as expected.

Change-Id: Ie9a6a5d48249373b05387e344643d1d3d1dbc8d2