[wlan] Fix ownership of wlan_tx_packet.packet_head
In all instances where the wlan_tx_packet struct is used, its
packet_head member (of type ethmac_netbuf_t) is assumed to be valid.
Enforce this by making this a struct member, instead of a pointer to an
external allocation.

This also fixes a case in wlan::Packet::AsWlanTxPacket() where a pointer
to a stack-allocated ethmac_netbuf_t was being returned in a member
pointer, leading to invalid memory access.

Test: Built and run the WLAN stack on a Pixelbook.

Change-Id: I75248a627cdad47973fe55014555a7a0d3d24abd