[wlan][ralink][ath10k] Update radio capability reports
Update the hardware capabilities reported by Ralink and Ath10k
with the ones that are experimentally proven-to-work. The experiments
were performed for more than several months by MLME's
BuiltHtCapabilities() overrider.

The updated values do not necessary reflect the true capabilities
of the underlying hardware. They need to be yet discovered and
time-tested before they can be adopted.

Test: Build and Smoke
NET-1321

Change-Id: Ia32f7e34965724c97fc5ce6536e3e7e36937126d