[auth] Expand store to include private key.
This is a preperatory step for IoT ID integration, which requires a
credential key with for each bound refresh token.

We also refactor a couple of the database fields for better consistency
with the rest of the codebase and to reflect the fact the same IdP
account may have different credentials generated by different
AuthProviders.

AUTH-54

Tested: CQ, including a couple of new unittestsa
Change-Id: I180af3f9a6f239c9e9e521ed0adbde8472ec900c