[ledger] Handle cloud provider disconnection / shut down.
This patch allows Ledger to continue to work when cloud provider
disconnects - for example, because it crashed, or because the framework
failed to initialize it because we are offline and the token manager
couldn't retrieve the Firebase user ID.

In the next step we should implement the proper recovery path - so that
we listen to the connectivity API and request a new cloud provider
instance when the connectivity is back.

In the meantime, this patch allows to continue working without cloud
sync.

LE-581 #done
LE-576 #done

Test: e2e test, unittests
Change-Id: I31a94c7a41cff76f4a395586f0c5724283d54755