[build] provide ENABLE_INSTALL_SAMPLES build parameter
When true, modules marked as sample or ddk-sample will be included
in the zircon.zbi.  Otherwise they will not be included.  They will
always be built.

Previously there was no way to override USER_MANIFEST_GROUPS,
and if you did (in engine.mk) change it to the empty string,
the --groups= that got passed to zbi would filter out *everything*.

Tested: manually

Change-Id: Ie975666c7817e020908f1681fab13f987e8783a7