[Camera] Introducing the Camera Manager
The Camera Manager Service offers a way for applications
to discover and connect to cameras and other input devices.
The CameraDeviceManager implements this service, by listening
for devices as they are created and populating an internal list.
When devices connect to a camera or video input, they receive
a video stream, which is a direct interface to the video input
driver.

TEST = CQ

Change-Id: I574258386a4caa5a8a1ae93233161ba5489eaf3c