[binding_set] Make sure binding is destroyed after it is deleted from the list of bindings
Motivation is that in
https://fuchsia-review.git.corp.google.com/c/peridot/+/186208, I want to
treat binding_set as the source of truth of which channels are active.

When a channel is closed, I want to know which channels are remaining,
so I hooked it up a callback to the destructor of the binded impl which
looks up the binding set. The problem is that the binding set contains
nullptr because the binding hasn't been fully removed yet.

TESTED: unit test, paved, tested against aforementioned CL
Change-Id: I0562e73b05359abf85ea6e41f47e63bce84ffb64