[utest][threads] Fix a race condition in thread-test
When running this test with a single CPU (by passing "-s 1" to
run-zircon) and as part of core-tests (by passing "-c
userboot=bin/core-tests"), I saw the following failure in
test_suspend_multiple:

        system/utest/core/threads/threads.cpp:708:bool test_suspend_multiple():
        :
        Comparison failed: info.state == ZX_THREAD_STATE_BLOCKED_EXCEPTION is false
        Specifically, 1 (0x1) == 259 (0x103) is false

The thread was still in the state ZX_THREAD_STATE_RUNNING (1).  This
failed because it is racy to use zx_object_get_info() to assert the
blocked state of a thread without polling for this blocked state.  Do
the latter by using wait_thread_blocked(), as other tests do.

Bug: none
Test: "run-zircon -a x64 -s 1 -c userboot=bin/core-tests"
Change-Id: I4606d97799d8c9e373d50880c289698b828472ad