[mediaplayer] introduce PayloadBuffer
This CL introduces PayloadBuffer, a new class used to manage payload
buffers in the player. This change is needed to support the new
payload allocation scheme, which will allow, among other things, the
video decoder and video renderer to share the VMOs they use to store
decoded video frames.

PayloadBuffer is used instead of the size/pointer combo that was
previously associated with the Packet class. PayloadBuffers are
managed using fbl::unique_ptr (though fbl::RefPtr could be used in
the future if that's needed). PayloadBuffer exploits the 'recycling'
scheme of fbl::unique_ptr to do cleanup when PayloadBuffers are
deleted. They also allow recyclers to be nested so that auxilliary
cleanup (e.g. disposing of FFMPEG resources) can be performed.

The recycling approach gets rid of the need for special subclasses
of Packet, which were used previously for buffer cleanup.

TEST: Tested for regression using stress tests and CQ.

Change-Id: I09346674493b4172e1000907b1f69cc0ef99fd8f