[virtcon] Define missing symbols needed by vc-input.cpp
vc-input.cpp depends on vc_toggle_framebuffer and g_vc_owns_display that
are defined by vc-display.cpp. However, vc-display.cpp is not being
included when building virtual-console-test. This wasn't an issue in the
past because --gc-sections would discard all functions that invoke these
symbols, but this breaks with profiling which is incompatible with
--gc-sections.

Test: manually
Change-Id: Ia6ab73b3599d996957ae3eb63f11e788ca96bbfb