[auth] Add new token manager factory to device_runner.
This cl creates a handle to the new token manager factory interface in device
runner and provides it to user provider for generating different instances of
token manager for each user.

All the user impls (user_provider, user_controller and user_runner) are
modified to take both the interfaces for the interim until we deprecate the old
one. Instead of passing token manager factory, token_manager handle is
directly passed into user_runner which is scoped per user and per app (in
this case, it will be framework's device_runner_app).

This cl doesn't provide any implementation yet for the new interface, and will
be added in subsequent small cls.

TEST=Manually tested the existing login flow, CQ

AUTh-81 #comment

Change-Id: Ic62118ca895e95c553abdcc40ab73d504c471b41