[netstack] inscrease the arena numBuffers to 2048
NET-486 #comment
The shared arena's numBuffers is 1024 and the fifo size of each eth
interface is 256 (bufs), which casues problems with a 5-eth-interface
device, as 256 * 5 > 1024. The arena (free bufs) is not big enough
to serve the 5 eth interfaces simultaneouly.
So in this CL, the numBffers is increased from 1024 to 2048.

Test:
-- Manully tested with a 5-eth-interface device. The ports with
   links up are able to get valid IP address through DHCP
   service successfully.
   (DHCP requests are using arena to transmit data.)

Together with next CL (to be created) and
https://fuchsia-review.googlesource.com/c/garnet/+/186577,
to fix NET-486. i(With a multi-port device,
the plugged ethernet port does not get a valid IP addr through
DHCP.)

Change-Id: Idc4c091b23c95a406c408592ad0a57a41372d163