[debugger] Improve value formatting.
Most of the change is internal in the structure of how the value
formatter works. All formatting functions are now on a class which can
manage the callbacks in one place.

Support is now added for structs and arrays with things like nested
strings (previously anything requiring fetching memory wouldn't print in
a nested context).

All lambdas are now removed except where the code calls out to other
asynchronous systems. The ValueFormat class maintains a hierarchy of
pending formatting to manage all pending work.

Calling OutputBuffer::WithContents was getting tiring and not adding
clarity, so the OutputBuffer constructor now supports creation with an
initial string directly. Other callers will be updated in a followup.

Slight syntax highlighting improvements are made: errors and types for
pointers are printed with a "comment" syntax which makes the output
easier to scan. Backtrace printing duplicated the frame number, and
the inverse video (which I thought looked weird) is now changed to
"important" syntax which looks nicer. The object index for other
listings is changed for consistency.

A minor bug in the ConsoleContext is fixed with frame numbers being
invalidated if it changes.

TEST=unit

Change-Id: I512d33db133ec7224ced7c709e18477892217573